requirements.txt
python-dotenv
openai
termcolor
langchain
langchain-openai
pyaudio
picamera
pygame